requirements.txt
Django==1.9.4
djangorestframework==3.3.3
appdirs==1.4.3
Confluence-py==1.0.1
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
django-cors==0.1
django-cors-headers==2.0.2
django-secure==1.0.1
django-sslserver==0.19
django-toolbelt==0.0.1
gunicorn==19.7.1
jira==1.0.10
MySQL-python==1.2.5
oauthlib==2.0.2
ordereddict==1.1
packaging==16.8
pbr==2.1.0
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.12.1
requests-oauthlib==0.8.0
requests-toolbelt==0.7.1
six==1.10.0
static3==0.7.0
xml2json==1.1
pymysql==0.7.11